# lms-server/src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: LMS API
  description: |
    Multi-tenant Learning Management System API
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Multi-Tenancy
    All resources are scoped by tenant. The tenant ID is automatically extracted from:
    - `x-tenant-id` header, or
    - The authenticated user's tenant
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Auth endpoints: 5 requests per 15 minutes
    - Password reset: 3 requests per hour
    - Tenant signup: 2 requests per hour
  version: 1.0.0
  contact:
    name: API Support
    email: edinmavric10@gmail.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api
    description: Development server
  - url: https://api.example.com/api
    description: Production server (example)
...
paths:
...
  /lessons:
    get:
      tags:
        - Lessons
      summary: Get all lessons
      operationId: getAllLessons
      security:
        - bearerAuth: []
      parameters:
        - name: course
          in: query
          schema:
            type: string
          description: Filter by course ID
        - name: title
          in: query
          schema:
            type: string
          description: Filter by lesson title (case-insensitive partial match)
        - name: date
          in: query
          schema:
            type: string
            format: date
          description: Filter lessons by date
      responses:
        '200':
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Lesson'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags:
        - Lessons
      summary: Create lesson
      description: Create new lesson with date and time (admin or professor)
      operationId: createLesson
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - course
                - title
                - date
                - startTime
                - endTime
              properties:
                course:
                  type: string
                  description: Course ID
                title:
                  type: string
                  description: Lesson title
                content:
                  type: string
                  description: Lesson content (optional)
                materials:
                  type: array
                  items:
                    $ref: '#/components/schemas/LessonMaterial'
                  description: Lesson materials (optional)
                date:
                  type: string
                  format: date
                  description: Lesson date
                startTime:
                  type: string
                  pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
                  example: '09:00'
                  description: Lesson start time in HH:mm format
                endTime:
                  type: string
                  pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
                  example: '11:00'
                  description: Lesson end time in HH:mm format (must be after startTime)
      responses:
        '201':
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /lessons/{id}:
    get:
      tags:
        - Lessons
      summary: Get lesson by ID
      operationId: getLessonById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Lessons
      summary: Update lesson
      description: Update lesson (only course professor or admin)
      operationId: updateLesson
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LessonId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Lesson title
                content:
                  type: string
                  description: Lesson content
                materials:
                  type: array
                  items:
                    $ref: '#/components/schemas/LessonMaterial'
                  description: Lesson materials
                date:
                  type: string
                  format: date
                  description: Lesson date
                startTime:
                  type: string
                  pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
                  example: '09:00'
                  description: Lesson start time in HH:mm format
                endTime:
                  type: string
                  pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
                  example: '11:00'
                  description: Lesson end time in HH:mm format (must be after startTime)
      responses:
        '200':
          description: Lesson updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Lesson'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Lessons
      summary: Delete lesson
      description: Delete lesson (only course professor or admin)
      operationId: deleteLesson
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Lesson deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  parameters:
...
    LessonId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Lesson ID
...
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    Lesson:
      type: object
      properties:
        _id:
          type: string
        tenant:
          type: string
        course:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                name:
                  type: string
                professor:
                  oneOf:
                    - type: string
                    - type: object
                      properties:
                        _id:
                          type: string
                        firstName:
                          type: string
                        lastName:
                          type: string
                        email:
                          type: string
        title:
          type: string
        content:
          type: string
        materials:
          type: array
          items:
            $ref: '#/components/schemas/LessonMaterial'
        date:
          type: string
          format: date
          description: Lesson date
        startTime:
          type: string
          pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
          example: '09:00'
          description: Lesson start time in HH:mm format
        endTime:
          type: string
          pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
          example: '11:00'
          description: Lesson end time in HH:mm format
        isDeleted:
          type: boolean
        createdBy:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
        updatedBy:
          oneOf:
            - type: string
            - type: object
              properties:
                _id:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
        createdAt:
          type: string
          format: date-time
    LessonMaterial:
      type: object
      required:
        - type
        - url
      properties:
        type:
          type: string
          enum: [pdf, video, presentation, link]
        url:
          type: string
        storageKey:
          type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...